refactor: 데이터 관리 객체를 Dao 로 변경 및 repository를 domain 영역으로 옮김
