+++ app/src/main/java/woowacourse/shopping/presentation/productdetail/ProductDetailActivity.kt
@@ -8,7 +8,7 @@ import android.view.MenuItem
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
-import woowacourse.shopping.data.cart.CartDbAdapter
+import woowacourse.shopping.data.cart.CartDao
 import woowacourse.shopping.data.cart.CartDbHelper
 import woowacourse.shopping.databinding.ActivityProductDetailBinding
 import woowacourse.shopping.presentation.model.ProductModel
@@ -65,7 +65,7 @@ class ProductDetailActivity : AppCompatActivity(), ProductDetailContract.View {
         intent.getParcelableExtraCompat<ProductModel>(PRODUCT_KEY_VALUE)?.let {
             presenter = ProductDetailPresenter(
                 this,
-                CartDbAdapter(CartDbHelper(this)),
+                CartDao(CartDbHelper(this)),
                 it,
             )
         } ?: noIntentExceptionHandler(getString(R.string.product_model_null_error_message))
